Split Vestibule into MediaPreview and VestibulePanel
