+++ src/components/Vestibule/MediaPreview/MediaPreview.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import './MediaPreview.css';
+
+function MediaPreview() {
+
+  return (
+    <div className="media-preview">
+      <div className="media-preview-placeholder"></div>
+    </div>
+  );
+}
+
+export default MediaPreview;
