+++ src/components/Vestibule/MediaPreview/index.js
@@ -1,0 +1,2 @@
+import MediaPreview from './MediaPreview';
+export default MediaPreview;
