+++ src/components/Vestibule/Vestibule.css
@@ -1,0 +1,7 @@
+.vestibule-container {
+  width: 100vw;
+
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+}
