+++ src/components/Vestibule/VestibulePanel/index.js
@@ -1,0 +1,2 @@
+import VestibulePanel from './VestibulePanel';
+export default VestibulePanel;
